src_ai/requirements.txt
openai
python-dotenv
pandas
scikit-learn
httpx
numpy
